pyproject.toml
[tool.poetry]
name = "term-paper-zaberov-django-6"
version = "0.1.0"
description = ""
authors = ["zaberov.dv"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.8"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
ipython = "^8.20.0"
crispy-bootstrap5 = "^2023.10"
pillow = "^10.2.0"
django-forms-bootstrap = "^3.1.0"
pytils = "^0.4.1"
django-crispy-forms = "^2.1"
redis = "^5.0.1"
celery = "^5.3.6"
django-extensions = "^3.2.3"
django-celery-beat = "^2.5.0"
eventlet = "^0.35.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
